fix(DeepwitherCommand): DeepwitherCommand.java の型キャストを修正
- DeepwitherCommand.java において、args[1] のパース処理で Long.parseLong を Double.parseDouble に修正しました。

+++ src/main/java/com/lunar_prototype/deepwither/command/DeepwitherCommand.java
@@ -96,7 +96,7 @@ public class DeepwitherCommand implements CommandExecutor, TabCompleter {
         }
 
         try {
-            double colorId = Long.parseLong(args[1]);
+            double colorId = Double.parseDouble(args[1]);
             double urgency = Double.parseDouble(args[2]);
 
             // メッセージの結合
